export styles into a separate file
